Rearranged test structure to split up no network and network tests
