+++ platformio.ini
@@ -9,7 +9,7 @@
 ; http://docs.platformio.org/page/projectconf.html
 
 [platformio]
-default_envs = quake
+default_envs = test_quake_no_network
 
 [env:teensy]
 platform = teensy
@@ -51,7 +51,7 @@ lib_ignore = ChRt, rwmutex, Drivers
 build_flags = -std=c++14 -Wall -D UNITY_INCLUDE_DOUBLE -fexceptions
 src_filter = +<native.cpp>
 
-[env:quake]
+[env:test_quake_no_network]
 platform = teensy
 board = teensy35
 framework = arduino
@@ -60,4 +60,15 @@ src_filter = -<*>
 test_build_project_src = true
 lib_ignore = Piksi, SpikeAndHold
 upload_protocol = teensy-cli
-test_filter = test_quake
+test_filter = test_quake_no_network, test_quake
+
+[env:test_quake_network]
+platform = teensy
+board = teensy35
+framework = arduino
+build_flags = -std=c++14 -Wall -D UNITY_INCLUDE_DOUBLE
+src_filter = -<*>
+test_build_project_src = true
+lib_ignore = Piksi, SpikeAndHold
+upload_protocol = teensy-cli
+test_filter = test_quake_network, test_quake
